Fix shell unit tests to work in linux environment
Modifies the shell unit tests so that we are only looking for the `sh`
specific command prompt of '$' in the output and not a more specific
string like `sh-3.2$` which appears in the terminal output on osx but
not linux. Also modifies the TestReplay test to compare replay outout to
previous output from stdout instead of checking against hard-coded
strings/lengths. This makes the test independent of the exaxt terminal
output which may very slightly different depending on the host OS