CWC-1654: Add exponential backoff to websocket.Connect and split it up a bit (#54)
The problem (from Jira): Our websocket retry logic is flawed if there is a transient network error. The reason is we do try to reconnect immediately, and if we fail to reconnect more than 2 times in a row , the code currently will not retry.

The solution (from the docstring for Connect):
In order for Connect to serve as a robust abstraction for other processes that rely on it, it must handle its own retry logic. For this, we use an exponential backoff. Some failures within the connection process are considered transient, and thus trigger a retry. Others are considered fatal, and return an error

I also took the liberty of reorganizing the code of Connect slightly to make it easier to read. Totally happy to negotiate on that, since it's arguably a matter of taste, but I personally found it hard to reason about a function that long.