Send signalR invocation messages with an invocationId so we can correlate completion messages and log errors
This refactors our signalR websocket code to send invocation messages
with a unique id so that the server can send a completion response for
that invocation. When we get the response we log that it was completed
successfully or if an error occurred we log the error message. Responses
are returned in a custom WebsocketResponse json object which currently
only contains a boolean indicating if the method was successfully and an
error message in case there was an error.