Adds shell plugin support to the daemon
This adds a new shell plugin to the daemon which implements the existing
shell websocket keysplitting protocol as shell-connection-service did in
typescript. It functions differently then the other plugins in that it
is not meant to be run as a long-living background process. Instead when
started with the shell plugin the daemon is expected to inherit the zli
process' std in/out. The daemon then handles reading/writing directly to
std in/out. When the shell daemon exits (either due to protocol error or
from a shell/quit stream message that indicates the pty was killed on
the target) the process is directly killed using os.exit(). Meanwhile
the zli just needs to spawn the daemon as a subprocess, let it do all
the work, and wait for it to exit before exiting the zli command.