Add attach support for shell daemon plugin
+ Adds an additional daemon command line flag for dataChannelId and if the
daemon is started with this flag set and plugin=shell it tries to attach
to an existing data channel instead of opening a new data channel.

+ UnixShell action is now started with attach flag. If its started with
attach=true then it will send a ShellReplayMessage as the first data
message in the websocket (agent should already have pty created but
should replay the output buffer from the pty). Otherwise if attach=false
it will send a ShellOpenMessage as the first message in the websocket
(agent should create a new pty).