[Agent] Fix websocket connections not closed properly (#84)
* [Agent] Fixes a bug where websocket connections were not properly closed when receiving messages in control channel

+ This fixes a bug where we were not properly closing the websocket
when receiving a CloseWebsocket message in the control channel. This
also fixes the websocket tmb logic to track both send/receive go
routines.

+ Also fixed a bug in defaultshell where we would not kill the pty
command process. This adds extra insurance on top of just closing the
pty fd that the default shell plugin will be killed properly as the
the pseudoterminal done channel waits for the pty command to complete.