Revert change of using Original()
It's technically not a bug that we're using parsed/cleaned version instead of original string. As long as we use the same version that the agent does, we're okay.

If we want to use Original(), we must also use Original() in agent.
Reverting this change because agent still uses String().

I originally made this change because unit test was failing, but
realizing now that my test was asserting the wrong thing, and should
assert using String() just like the daemon and agent build their
messages.