Split IBZCert interface methods that are specific to daemon into IDaemonBZCert
+ Adds a new VerifyAndExitOnError method which will verify the bzcert is
valid and will call OS.exit with custom error code and log errors in the
case of invalid/expired IdP token errors.

+ Removes Refresh() call from BuildSyn so we dont try and refresh
everytime and instead only Refresh in the Recover method of keysplitting
when we get a KeysplittingError. If refreshing succeeds also call
VerifyAndExitOnError again to ensure the BZCert is valid and exit early
if its not.

+ Fixes some of the keysplitting tests to use common BeforeEach method
when setting up mocks/sut with defaults. This required moving to pointer
based interface methods as well as clearing the mocks and explicitly
setting all mocked functions for tests which needed specific mock
values.