Fix race condition in new data channel flow (#112)
* Refactors signalR websocket method selector logic to be internal to websocket.go

Instead of passing down a lambda to select the websocket method to call
websocket.go now determines this based on the type of websocket. This
also adds a new type SignalRWebsocketMethod and constants for all
websocket method names.

* Refactor websocket.go to allow deserialization of other types of websocket payloads

Our websocket code was assuming an signalR invocation call would have an
AgentMessage as an argument payload. This made it impossible to
send/deserialize other types of payloads from the connection node. Now
we partially deserialize the signalR invocation method to determine the
method being called and then deserialize the specific payload based on
the target method in a second step.

* Wait for AgentConnected message before sending websocket output messages in the Daemon

This fixes a race condition where we would sometimes send the
openDataChannel message before the agent websocket had connected and the
agent was just dropping the message